Add sys.hash_info and update number hashing

+++ Src/IronPython/Lib/iptest/type_util.py
@@ -5,7 +5,7 @@
 # types derived from built-in types
 
 class myint(int): pass
-class mylong(long): pass
+class mylong(int): pass
 class myfloat(float): pass
 class mycomplex(complex): pass
